requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0

streamlit==1.35.0

langchain==0.2.0
langchain-groq
openai==1.30.1
groq==0.5.0

python-dotenv==1.0.1
pydub==0.25.1
soundfile==0.12.1
SpeechRecognition==3.10.1

requests==2.32.3
httpx==0.27.0

